Allow running chef-marketplace-ctl register-node as a non-root user
Run the embedded chef-client run with a globally accessible cache
so that unprivileged users can run the command.  This will allow
the chef-compliance core application to run the command for us
during the setup wizard.  It should also allow the opscode user
to do the same if we implement registration in chef-manage.