mainTemplate.json - set default vnet address prefix
When using an existing virtual network in the Azure UI, the
parameters passed to the ARM template do not include the
vnet address prefix because it's already defined as part of
the existing vnet. This causes the ARM template validation to
fail, even though it's a non-required parameter for this
action.

By setting a default value, we can bypass the case where the
address prefix is not supplied because we truly do not need
it.