[CLOUD-291] Add Automate ARM template matrix test
The Azure Marketplace publishing guidelines require that each Solution
Template that we submit to Azure be deployed at least 30 times. This
commit adds a testing utility with 30+ likely permutation scenarios.
After the testing has completed the logs of each test scenario are
aggregated into a zip archive which we submit to Microsoft when we
request that they publish our staged solution template.