[CLOUD-269] Disable Biscotti during Automate Setup
During the setup there's a short period of time where Automate's
nginx will be able to successfully proxy to Biscotti's setup page
before Chef Automate and Chef Server have been fully configured.
This is a major problem if the user submits the setup form
and Biscotti attempts to create accounts and a starter kit without
a fully functional cluster. This change disabled the Biscotti
service while the cluster is being configured to ensure that the
setup page is not available until it can be safely used.