[CLOUD-371] Fix pedant tests for Chef Automate AIO
This change overrides the Chef server pedant configuration exclude the
Chef Server's non-default SSL port. That will allow the tests to run
through the Chef Automate proxy and into the Chef Server.

It also limits SSL to TLSv1.2, which is fine because SSL is terminated
on the Chef Automate nginx. It also fixes a bug where some pedant tests
would attempt to use TLSv1.0 against the Automate nginx and it was
rejected.