Disable fail_timeout and max_fails in nginx
We are seeing issues where biscotti will return with a server side
error before the backend retries can complete. This disables the
fail_timeout and max_fails for the upstream directive to ensure we
always retry an upstream evcen if it failed to return a good HTTP
response in case the unicorn master nukes a single worker for timing
out.

Signed-off-by: Ryan Hass <rhass@users.noreply.github.com>