Add debug option to stream stderr/stdout.
This allows us to see errors when a failure occurs in realtime.
Without this, we only have basic knowledge of a failure without
details as to why it may have occurred. This should NOT be used
as a defalt parameter for creating non-developer systems as
sensitive data can get exposed such as private keys and passwords.

Signed-off-by: Ryan Hass <rhass@users.noreply.github.com>