Use shelljoin on command strings.
The `delivery-ctl create-enterprise` step is failing when the command
is executed. The output from the failure shows the command having been
executed without escaped characters, while the printed error message
includes them. This suggests the `retry_command` is not actually
passing the correct string to be executed. Rather than escaping just a
partial command, this is inteded to escape the entire command string to
prevent mixlib-shellout from doing something other than what we
actually intend.

Signed-off-by: Ryan Hass <rhass@users.noreply.github.com>