Set HOME environment variable on initial startup.
`automate-ctl create-enterprise` is called via the marketplace gem and
chef-marketplace-ctl. We execute the chef-marketplace-ctl commands
automatically on first boot with the `--preconfigure` flag to configure
the Chef Automate and Chef Server products, as well as allow users to
configure their credentials via the biscotti interface. The
`chef-marketplace-ctl` command calls `automate-ctl`, which calls an escript
`enterprise_ctl` which requires the environment variable `HOME` to be
defined as a valid path. This ensures we set a value and allows the
enterprise to be created.

Fixes SUSTAIN-844 #comment Set $HOME environment variable to prevent
`delivery-ctl create-enterprise default` from failing.

Signed-off-by: Ryan Hass <rhass@users.noreply.github.com>