Point ARM template to older revision
The ARM test suite is failing and it is unclear from the resulting logs
is this was caused by the changes made to the ARM templates to fix the
HOME environment variable, or if there is some other issue. The
`azure-baseline` branch is prior to any changes made by me with the
mainTemplate.json updated to point to it for baseline testing.

Signed-off-by: Ryan Hass <rhass@users.noreply.github.com>