Enable Authorization on Azure.
This enables the Authorization form on Azure instances to ensure the
person configuring the first set of credentials is indeed the owner of
the instance. This also fixes some bugs in which the biscotti setup
page would not load as the nessicary tokens were not set by the
Authorization angular app befor they were attempted to be used by the
backend biscotti app.

Signed-off-by: Ryan Hass <rhass@users.noreply.github.com>