AppBundle reckoner application & add activesupport
The reckoner application is failing to start correctly in the lastest
metered AMI submitted to AWS due to pathing issues and inability to
load specific gems. This resolves the issue by using appbundler to
explicitly set the gem path and pin the gem versions at runtime.

Additionally, we made use of active_support/time numeric durations
without having included activesupport in the gemfile or requiring it
into script resulting in undefined method errors. This fixes the calls
so they can function as originally intended.

Signed-off-by: Ryan Hass <rhass@users.noreply.github.com>