Only depend on aws-sdk-marketplacemetering and update omnibus gem
A new omnibus gem was recently released that fixes
a bug in the omnibus buildkite pipelines which requires
updating the omnibus gem in this repo's Gemfile.lock file.
However, the current Gemfile depends on an old version of aws-sdk
which conflicts with the omnibus' use of the newer aws-sdk version
that breaks aws-sdk into individual gems to enable user's to
avoid depending on hundreds of unnecessary aws-sdk gems.
It looks like omnibus-marketplace reckoner only needs
the aws-sdk-marketplacemetering gem.

Signed-off-by: Jeremiah Snapp <jeremiah@chef.io>